Fix: pskb-copy segment with correct head-len
pskb_copy just copies the necessary space for THIS segment. However,
when we are reinjecting we do not yet know the necessary header-space.
E.g., SACK, REMOVE_ADDR,... may need to get added to the header-space.
In this case, the header-space would not be sufficient.

Thus, do like tcp_retransmit_skb, and take the maximum header space.

Signed-off-by: Christoph Paasch <christoph.paasch@uclouvain.be>
Signed-off-by: Gregory Detal <gregory.detal@uclouvain.be>