Fix: Correctly close subflows
A subflow is being closed only through a call to tcp_close, thus setting
tp->closing to 1. It may happen, that a subflow is not being closed,
while the meta-socket is dying. E.g., upon mptcp_sub_close, the meta had
not yet sk_shutdown et to SHUTDOWN_MASK. Thus, only a subflow-fin will
be sent and the sk_state of the subflow changed.

We must make sure that *all* subflows get closed properly when the meta
is dying. This is achieved through checking the tp-closing flag in
mptcp_destroy_sock.

Further, this patch correctly sets sk_shutdown when we only send a
subflow-fin on a subflow. Finally, we call more agressively
mptcp_sub_force_close(), as this will make sure that the subflows really
disappear.

Signed-off-by: Christoph Paasch <christoph.paasch@uclouvain.be>